Deadline commit
Optimized imaging_data load, tried to do it to scan frame, but not succeeded. Anyway, if scan_frame is slow, that is because of the ReadAnalogF64 function + machine_config AIO sample rate + hardware limitations, not the code.
Finished Z stack and made it uninterruptibe (for the simplicity). It can be controlled via ZMQ too. ZMQ support added to most of the functions (especially, which is made to be controlled remotely).
Minor bug fix in generate_waveform (removed last sample from self.waveform).
Scan restarts only if waveform parameters getting changed.
Added some comments in order to help development in the future.
Parts still missing: 
 - get camera image (hardware was not specified, could not write code about it)
 - binning
 - set depth (hardware unknown)